pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "irtoolshed-mcp-server"
version = "0.1.0"
description = "MCP Server providing tools for network incident response"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cymruwhois>=1.6",
    "mcp[cli]>=1.6.0",
    "dnspython>=2.6.0",
    "python-whois>=0.8.0",
    "geoip2>=4.8.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "black>=24.2.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=irtoolshed_mcp_server"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["irtoolshed_mcp_server"]
